Add module for implementing 'finally' in Lua.
`finally` takes two functions, `pcall`s the first one and makes sure
that the second function is called after the first regardless of whether
the first raised an error or not. The C version takes special care to
preallocate memory so that execution of the second function cannot be
interrupted.