Avoid possible memory allocation error on Lua 5.1.
`lua_checkstack()` can cause a memory allocation error on Lua 5.1, so
there is no safe way to preallocate a given amount of stack slots so
that it will survive garbage collection until the cleanup function runs.
As a workaround the Lua closure responsible for allocating call frames
also allocates a few extra locals.